requirements.txt
astroid==2.4.2
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
folium==0.12.1
geographiclib==1.50
geopy==2.1.0
idna==2.10
IMDbPY==2020.9.25
isort==5.7.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.20.1
pandas==1.2.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
urllib3==1.26.3
wrapt==1.12.1
